Server JWT/requirements.txt
Flask==0.10.1
hashids==1.1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PyJWT==1.4.0
requests==2.8.1
Werkzeug==0.10.4
wheel==0.24.0
